+++ src/tweens/cubic.rs
@@ -5,6 +5,9 @@ declare_tween!(
     /// Creates a new [CubicIn] Tweener.
     pub fn cubic_in;
 
+    /// Creates a new [CubicIn] Tweener at the given time.
+    pub fn cubic_in_at;
+
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, percent: f32) -> Value {
         value_delta.scale(percent * percent * percent)
     }
@@ -16,6 +19,9 @@ declare_tween!(
 
     /// Creates a new [CubicOut] Tweener.
     pub fn cubic_out;
+
+    /// Creates a new [CubicOut] Tweener at the given time.
+    pub fn cubic_out_at;
 
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, mut percent: f32) -> Value {
         percent -= 1.0;
@@ -30,6 +36,9 @@ declare_tween!(
 
     /// Creates a new [CubicInOut] Tweener.
     pub fn cubic_in_out;
+
+    /// Creates a new [CubicInOut] Tweener at the given time.
+    pub fn cubic_in_out_at;
 
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, mut percent: f32) -> Value {
         percent *= 2.0;