+++ src/tweens/quint.rs
@@ -5,6 +5,9 @@ declare_tween!(
     /// Creates a new [QuintInOut] Tweener.
     pub fn quint_in;
 
+    /// Creates a new [QuintInOut] Tweener at the given time.
+    pub fn quint_in_at;
+
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, percent: f32) -> Value {
         value_delta.scale(percent * percent * percent * percent * percent)
     }
@@ -16,6 +19,9 @@ declare_tween!(
 
     /// Creates a new [QuintOut] Tweener.
     pub fn quint_out;
+
+    /// Creates a new [QuintOut] Tweener at the given time.
+    pub fn quint_out_at;
 
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, mut percent: f32) -> Value {
         percent -= 1.0;
@@ -29,6 +35,9 @@ declare_tween!(
 
     /// Creates a new [QuintInOut] Tweener.
     pub fn quint_in_out;
+
+    /// Creates a new [QuintInOut] Tweener at the given time.
+    pub fn quint_in_out_at;
 
     pub fn tween<Value: crate::TweenValue>(&mut self, value_delta: Value, mut percent: f32) -> Value {
         percent *= 2.0;